requirements.txt
imutils==0.5.3
opencv-python==4.2.0.34
flask==1.1.2
dlib==19.19.0
pyyaml==5.3.1
pytest==5.4.1
pytest-mock==3.1.0
pytest-cov==2.8.1
numpy==1.18.3
